Make separate file for temp user data
